Extract cli module, collapse monolith to a compatibility shim (split PR 9/N, final)
Moves main() and get_args() into tfbs_footprinter3/cli.py, then
reduces tfbs_footprinter3.py itself to a pure re-export shim.

- tfbs_footprinter3/cli.py (new): main, get_args + the signal.signal()
  hook. Imports everything it orchestrates from the now-settled set of
  per-concern modules (alignment, data_loader, data_parsing, io_utils,
  output, pipeline, plotting, scoring, translators).
- tfbs_footprinter3/tfbs_footprinter3.py: shrinks from ~524 lines to
  112 lines of pure re-exports. No implementation here anymore; every
  symbol is forwarded from its owning module.
- Existing callers (tests, hpc/, downstream) that do
  'from tfbs_footprinter3.tfbs_footprinter3 import <anything>' are
  untouched — all names still resolve.
- The console-script entry point
  'tfbs_footprinter3 = tfbs_footprinter3.tfbs_footprinter3:main' still
  works (verified: 'main.__module__' == 'tfbs_footprinter3.cli',
  reached via the shim).
- Dead-code cleanup: drops the commented-out shadow definitions that
  lingered in main() (all_tfs_thresholds/all_pwms_loglikelihood_dict
  loading, the various 'alternative signature' comments on
  tfbs_finder/find_clusters/species_specific_data calls). All were
  superseded by their active successors already.

Final module sizes after the full split:
    alignment.py       113
    cli.py             384
    data_loader.py     310
    data_parsing.py    104
    ensembl.py         103
    io_utils.py         96
    output.py          124
    pipeline.py        365
    plotting.py        680
    pwm.py              56
    scoring.py         358
    translators.py     413
    tfbs_footprinter3.py (shim)  112
                              ----
                              ~3200 total (vs 3306-line original monolith)

All 21 unit tests green. ruff check clean. Monolith split complete.
